all the api code_29-08-2023

+++ main/views.py
@@ -10,7 +10,6 @@ from .segment  import *
 from .speakerFile  import *
 import os
 from rest_framework import status
-import pydub
 from pydub import AudioSegment
 
 from rest_framework.response import Response
@@ -256,25 +255,17 @@ def download_json(request, id):
 class Segmented_Text(APIView): 
     def post(self, request,format=None):
       serializer = AudioService(data = request.data)
-      print(serializer)
       language = request.POST.get('language')
       if serializer.is_valid():
           print(language)
           audio1 = request.FILES.get('Audio_File')
           file_take = AudioSegment.from_file(audio1)
-          sample_rate = file_take.frame_rate
-          samplerate = file_take.set_frame_rate(44100)
-          sample_width_bytes = file_take.sample_width
-          print("Sample Rate:", sample_rate, "Hz")
-          bits_per_sample = sample_width_bytes
-        #   bits_per_sample = sample_width_bytes * 8
-          print(f"Bits per sample: {bits_per_sample}")
+          samplerate = file_take.set_frame_rate(44100).set_sample_width(2).set_channels(2)
           file_path = f'C:\\Users\\91722\\Desktop\\git vala folder\\Whisper_TR\\media\\{audio1}'
           samplerate.export(file_path, format="wav")
           client = storage.Client(credentials=service_account.Credentials.from_service_account_file(r'C:\\tr_2\\main\\new.json'))
           bc_name = "avyaan_mgmt"
           bucket = client.bucket(bc_name)
-        #   audio_path = r"C:\\Users\\91722\\Desktop\\git vala folder\\Whisper_TR\\media\\audio\\" + str(audio1)
           audio_path = file_path
           blob = bucket.blob(audio_path)
           blob.upload_from_filename(audio_path)
@@ -288,7 +279,7 @@ class Segmented_Text(APIView):
           min_speaker_count=2,
           max_speaker_count=2,    )
           config = speech.RecognitionConfig(
-            encoding=speech.RecognitionConfig.AudioEncoding.ENCODING_UNSPECIFIED,
+            encoding=speech.RecognitionConfig.AudioEncoding.MP3,
             sample_rate_hertz=44100,
             audio_channel_count=2,
             language_code=language,
@@ -301,7 +292,7 @@ class Segmented_Text(APIView):
           start_time = None
           for result in response.results:
               for word_info in result.alternatives[0].words:
-                  if word_info.speaker_tag in [2, 1]:  # Only consider speaker 1 and speaker 2
+                  if word_info.speaker_tag in [1,2]:  # Only consider speaker 1 and speaker 2
                       if current_speaker_tag is None:
                         current_speaker_tag = word_info.speaker_tag
                         start_time = word_info.start_time
@@ -317,13 +308,74 @@ class Segmented_Text(APIView):
           for sentence, speaker_tag, start_time, end_time in sentences:
               final += " Speaker" + str(speaker_tag) +  " : "+ "Start Time: " + str(start_time) +  ' End Time: '+ str(end_time) +  " : " + sentence + "\n" 
           print(final)
-          bucket = client.bucket(bc_name)
-          file_blob = bucket.blob(audio_path)
-          file_blob.delete()   
-          print("deleted") 
+          try:
+              bucket = client.bucket(bc_name)
+              file_blob = bucket.blob(audio_path)
+              file_blob.delete()   
+              print("deleted") 
+          except Exception as e:
+              print(e)  
           serializer.save(name=str(audio1),description=final)
-          print("Success")
           return Response({"msg": final}, status=status.HTTP_201_CREATED)
+      return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+class No_Segmented_Text(APIView): 
+    def post(self, request,format=None):
+      serializer = AudioService(data = request.data)
+      print(serializer)
+      start_time = request.POST.get('start_time') 
+      end_time = request.POST.get('end_time')
+      print(start_time,end_time)    
+      language = request.POST.get('language')
+      if serializer.is_valid():
+          print(language)
+          print(type(start_time))
+          audio1 = request.FILES.get('Audio_File')
+          file_take = AudioSegment.from_file(audio1)
+          samplerate = file_take.set_frame_rate(44100).set_sample_width(2).set_channels(2)
+          z1 = int(start_time) * 1000
+          z2 = int(end_time) * 1000
+          print(z1,z2)
+          segment = samplerate[z1 : z2]
+          file_paths = r'C:\\Users\\91722\\Desktop\\git vala folder\\Whisper_TR\\media\\audio1'+str(audio1)
+          segment.export(file_paths, format="mp3")
+          client = storage.Client(credentials=service_account.Credentials.from_service_account_file(r'C:\\tr_2\\main\\new.json'))
+          bc_name = "avyaan_mgmt"
+          bucket = client.bucket(bc_name)
+          audio_path = file_paths
+          blob = bucket.blob(audio_path)
+          blob.upload_from_filename(audio_path)
+          gcs_uri = r'gs://' + bc_name + '/' + audio_path
+          audio = speech.RecognitionAudio(uri=gcs_uri)
+          client_file =r'C:\\tr_2\\main\\new.json'
+          credentials = service_account.Credentials.from_service_account_file(client_file)
+          client1 = speech.SpeechClient(credentials=credentials)
+          diarization_config = speech.SpeakerDiarizationConfig(
+          enable_speaker_diarization=True,
+          min_speaker_count=2,
+          max_speaker_count=2,)
+          config = speech.RecognitionConfig(
+            encoding=speech.RecognitionConfig.AudioEncoding.MP3,
+            sample_rate_hertz=44100,
+            audio_channel_count=2,
+            language_code=language,
+            diarization_config=diarization_config,)
+          response = client1.long_running_recognize(
+          config=config, audio=audio).result()
+          text = ""
+          for result in response.results:
+            for alternative in result.alternatives:
+                text += alternative.transcript
+          print(text)
+          try:
+              bucket = client.bucket(bc_name)
+              file_blob = bucket.blob(audio_path)
+              file_blob.delete()   
+              print("deleted") 
+          except Exception as e:
+              print(e)  
+          final = "Speaker: " + "Start_Time:- "+ str(start_time)+" End_Time:- "+ str(end_time)+" Text:- " + str(text)    
+          serializer.save(name=str(audio1),description=text)
+          return Response(final, status=status.HTTP_201_CREATED)
       return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
